+++ server/src/visitors/YmlParsingErrorListener.ts
@@ -1,7 +1,7 @@
 import { ANTLRErrorListener, RecognitionException, Recognizer, Token } from 'antlr4ts';
 import { Diagnostic, DiagnosticSeverity, Range } from 'vscode-languageserver';
 
-import { severity } from '../server';
+import { parsingIssueSeverityLevel } from '../server';
 
 export class YmlParsingErrorListener implements ANTLRErrorListener<Token> {
     constructor(public diagnostics: Diagnostic[]) {}
@@ -24,7 +24,7 @@ export class YmlParsingErrorListener implements ANTLRErrorListener<Token> {
         const diagnostic = Diagnostic.create(
             Range.create(currentEditorLine, charPositionInLine, currentEditorLine, endPosition),
             msg,
-            severity,
+            parsingIssueSeverityLevel,
         );
         this.diagnostics.push(diagnostic);
     }