+++ server/src/yml-objects/YmlMethod.ts
@@ -6,9 +6,4 @@ export class YmlMethod extends AbstractYmlFunction {
     constructor(indentifier: string, uri: string) {
         super(indentifier, CompletionItemKind.Method, uri);
     }
-
-    public getShortName() {
-        const functionNameSubParts = this.label.split('::');
-        return functionNameSubParts.pop();
-    }
 }