+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 1.6.5 (YYYY-MM-DD)
 
 -   Improved parsing for:
-    -   `forall` instructions
-    -   `(condition).check()`-like expressions.
+    -   `forall` instructions;
+    -   `(condition).check()`-like expressions;
+    -   `rename` instructions.
 
 ## 1.6.4 (2019-07-31)