+++ CHANGELOG.md
@@ -14,7 +14,7 @@
     -   parse `*.yclass`, `*.yobject` and `*.ycomplete` files;
     -   avoid parsing the files in `.generated-yml/` directory, wherever this directory stands.
 
-_Note that the whole project parsing may take some time, depending on your project size. During the project parsing, suggestions list won't be available._
+_Note that parsing a whole project may take some time, depending on the project’s size. During the project parsing, suggestion lists won't be available._
 
 ## 1.6.4 (2019-07-31)